Tone queue - from global variables to function argument.
Putting all global variables related to tone queue into single struct. The struct is still global, but some internal functions were modified to accept a 'cw_tone_queue_t' variable. So this is a step forward - passing tone queue as an argument instead of always referring to global variable.
One of struct's fields is mutex - tone queue is used in two different threads, so the mutex may be necessary.

The lines changed at the bottom of the file helped me prove that it is possible to use CW_USECS_* to generate desired wave, with desired time characteristics and slope (they are almost perfect, slope makes tones a bit longer than necessary. But this is a step in a not-so-bad direction.).
I will try to use this method in few other places, e.g. in straight key control.