Adding support for Null output.
Null output is an empty device that doesn't produce a sound. It doesn't use any physical or logical device to output a sound. Its only functionality is to provide a timing information: it acts as if it accepts and plays x microseconds of sound, and it returns after the x microseconds.
I'm not sure if this will turn out useful, maybe yes, maybe no.
Currently it is implemented using single call to usleep(), without any checks of return value. This should be improved.