Introducing "length" property of tone queue.
Adding tq->len field that will keep current length of queue (number of
elements in queue). It will be increased on successful enqueueing of a
tone, and decreased on successful dequeueing of tone.

At this point len is used in regular code and in test code. There are
two asserts comparing the len and capacity of tq that fail. I have
commented them out and marked with FIXME. Perhaps after more changes
in the code they will work as expected.

The len field is introduced to change behavour of capacity field. I
would like the capacity to store exact capacity of queue, not
"capacity + 1". I will need the len in order to have a correctly
working tq, and I will need to rework some functions using capacity,
head and tail.