Refactoring receiver code.
Small refactoring of receiver code. Graph of machine states for
receiver moved on top of "Receiving" section. Some code moved on top
of the section.

Few variables related to receiving moved to stuct called "receiver".
I plan to use the variable in the same way as "generator": to
conveniently hold all data related to receiving.