Adding more test points for fixed receive test.
cw_test_fixed_receive() contained only three characters to be tested:
Q, R, P. I've added all characters supported by libcw.

All the test points are held in a table. Timing parameters in the
table are for now without jitter, it will be added by a helper
function randomizing value of jitter.

Temporarily disabling one of tests that is not related to receiver,
but is time consuming.