Updating "tests/testedin" marks in comments.
A lot of functions gets a "tests::" or "testedin::" marks that
indicate which functions are tested in a given unit test function, or
in which unit test a given function is tested.  These marks are put in
top-level function's comment, and may be used (in case of
"testedin::") to quickly verify if a given function is tested
somewhere or not.