Adding definition of libtool version of library.
Previously the version has been defined in src/libcw/Makefile.am, but
I was unable to access the value in source code of library and
library's unit tests. After moving it to configure.ac and adding two
lines of code the version is now available both to build system (as it
was before) and to source code (through config.h).

Now unit test can look at the value defined in build system instead on
magic numbers entered in code of the unit test.