Changes related to fixing problem with clocking iambic keyer.
The core of changes is in argument passed to
cw_{start|stop}_receive_tone(). Before it was NULL, which meant that
the functions were generating the timestamp themselves. The timestamp
has been generated at not-so-precise intervals: whenever a
write-to-soundcard function has returned. Now the functions accept
timestamp generated in much more precise way. See change in libcw.c to
see how.

Other changes in the file are centered around this functions' argument
change. I hope that I've put enough comments to know what's going on
when I will be reading the code in two years from now.