Adding test_cw_rec_build_data()
Adding function that creates test data with mark/space timings. The data will be used to test receiver's input functions.
The calculated data should be almost the same as in libcw_test_public.c/TEST_DATA_RAW. It's generated by code, so it should be easier to verify its correctness and to modify it. The function also takes less space than the table.