Modifications to properly test is_word flag.
Receiver's poll() functions return is_word flag. Adding code that tests that the poll() functions correctly set the flag.
The flag is set to true whenever a space ending a character is longer than maximal length of end-of-character space. Any such space is by definition end-of-word space, which means that a receiver completed receiving a word.