Removing tests of receiver
All receiver tests are now stored in libcw_rec.c. Since the receiver
is independent from generator, tone queue or keyer, there is no need
to test receiver in 'public'. All tests (of public and internal
functions) can be performed within the module. The receiver test
functions are called in libcw_test_internal executable, which in turn
is built and executed when making 'make check' target.