Undoing some changes in cw_queue_tone() and cw_tq_enqueue_internal()
There should be a difference in what range of tone lengths is accepted by the functions. The _internal() function should be able to accept a special (negative) value of tone length. The special value is CW_AUDIO_FOREVER_USECS).

Also changing fprintf()s first argument to stderr, so that the string is not printed to console when invoking "make check".