Moving functions operating on global cw_rec to libcw.c
What is left in the file are _internal() functions accepting as a
first argument a pointer to cw_rec_t variable.
At this point no main function in this file is operating on the global
cw_receiver variable. The variable is still used by unit test code,
this still should be updated.