Getting rid of reference to global cw_receiver variable
Thanks to new field in cw_key_t it is now possible to remove 'extern cw_rec_t cw_receiver' and use key->rec instead. This means that cw_receiver can be now static in libcw.c.
Also adding function that binds key and receiver. It is unused at this moment, but since we have a function that binds key with generator...