Improving cw_key_ik_enqueue_symbol_internal()
Improving the cw_key_ik_enqueue_symbol_internal() function so that it
doesn't need to know anything about lengths of mark and space (which
are known to generator, but the key doesn't have to know about
them). The iambic keyer now only sends symbols (Space/Dot/Dash) to the
generator. Generator takes care of selecting proper length of symbol.