Unit tests
Adding declaration of unit test function.
The added unit test function uncovered one problem occuring when a generator is created, but not started, and then is deleted. A call to pthread_kill(&id, ...) on invalid id results in crash.
I had to add a new flag to cw_gen_t to fix this.