Making use of cw_gen_t->thread.running flag in cw_gen_silence_internal() to avoid segfaults
Testing the flag to avoid execution of the function when the generator's thread is not active anymore.
The execution (and pthread_kill() on invalid thread id in particular) would result in segfault.
I have discovered this on Alpine Linux - src/cw/cw crashed every time I pressed Ctrl-C.
I don't know how old this problem is. Maybe it has been introduced in this dev cycle, or maybe it was there since who knows when.