Fixing generator's consumer code that uses semaphores
Moving call to sem_wait() on top of while(gen->do_enqueue_and_play)
loop, before dequeue_internal() function. This means that the
dequeue_internal() will be executed only when a semaphore is posted,
i.e. when new tone is enqueued in tone queue (or when generator is
deleted). This is a nice improvement of program logic. Old calls to
sem_wait() in the loop are still in old places, but disabled.