Test statistics table improvements
Values in the table are now printed only for those tests (i.e. combinations of sound system and test topic) that were really tested. For non-tested combination nothing is printed in the table. Thanks to this the values showing statistics of performed tests are better visible: you don't have to scan visually the whole table to search for non-zero statistics.